some agg

+++ core/src/main/java/com/asml/apa/wta/core/model/Workflow.java
@@ -55,7 +55,7 @@ public class Workflow implements BaseTraceObject {
 
   private final long totalNetworkUsage;
 
-  private final long totalDiskSpaceUsage;
+  private final double totalDiskSpaceUsage;
 
   private final long totalEnergyConsumption;
